feat(update): add double precision and key controls

+++ frag.glsl
@@ -3,8 +3,8 @@
 layout(location = 0) in vec2 inPosition;
 
 layout(binding = 0) uniform RenderAreaInformation {
-    vec2 renderAreaSize;
-    vec2 renderAreaOffset;
+    dvec2 renderAreaSize;
+    dvec2 renderAreaOffset;
     uint numberOfIterations;
 };
 
@@ -12,13 +12,13 @@ layout(location = 0) out vec4 outColor;
 
 void main()
 {
-    vec2 c = vec2(renderAreaOffset.x + inPosition.x * renderAreaSize.x,
+    dvec2 c = dvec2(renderAreaOffset.x + inPosition.x * renderAreaSize.x,
             renderAreaOffset.y + inPosition.y * renderAreaSize.y);
-    vec2 z = vec2(0.0, 0.0);
+    dvec2 z = dvec2(0.0, 0.0);
     gl_FragDepth = 0;
     for (uint i = 0; i < numberOfIterations; i++)
     {
-        z = vec2(z.x * z.x - z.y * z.y, 2.0 * z.x * z.y) + c;
+        z = dvec2(z.x * z.x - z.y * z.y, 2.0 * z.x * z.y) + c;
         if (dot(z, z) > 4.0)
         {
             i = i % 12;
